auth.js 안 fetch 함수 fetch.js 로 이동 (20.11.03)
